Brooklyn-health/requirements.txt
asgiref==3.7.2
boto3==1.34.47
botocore==1.34.47
Django==4.2.10
jmespath==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
